+++ mg_string.h
@@ -18,12 +18,12 @@ struct string_ref {
     char* Ptr = nullptr;
     const char* PtrC;
   };
-  size_t Size = 0;
+  int Size = 0;
 
   string_ref() = default;
-  string_ref(cstr Ptr, size_t Size);
+  string_ref(cstr Ptr, int Size);
   string_ref(cstr Ptr);
-  char& operator[](size_t i);
+  char& operator[](int i);
   operator bool();
 }; // struct string_ref
 
@@ -34,13 +34,13 @@ bool operator==(string_ref Lhs, string_ref Rhs);
 
 /* Return a substring of a given string. The substring starts at Begin and has length Size.
 Return the empty string if no proper substring can be constructed (e.g. Begin >= Str.Size). */
-string_ref SubString(string_ref Str, size_t Begin, size_t Size);
+string_ref SubString(string_ref Str, int Begin, int Size);
 
 /* Tokenize strings without allocating memory */
 struct tokenizer {
   string_ref Input;
   string_ref Delims;
-  size_t Pos = 0;
+  int Pos = 0;
 
   tokenizer() = default;
   tokenizer(string_ref Input, string_ref Delims = " \n\t");
